// Proiect/BGJ-Shop-Admin/src/app/models/game.model.ts
export class GameModel {
  id: number;
  bggId: number;
  title: string;
  type: string;
  imageUrl: string;
  description: string;
  minPlayers: number;
  maxPlayers: number;
  playingTime: number;
  year: number;
  price: number;
  newPrice: number;
  stock: number;
  dateAdded: Date;
  dateModified: Date;
}